refactor: edited App.js to App.jsx
